requirements.txt
connexion
sqlalchemy
mysql-connector-python
pymysql
pyyaml
requests
uvicorn
connexion[flask]
apscheduler
swagger-ui-bundle
